refactor(httpd): update typing approach for fetchFromHttpd()
This is a less generic (was not extensible enough) and more
explicit typing approach which can also better document httpd API calls.

Additionally some file structure / renaming was done.
