+++ src/helpers/index.ts
@@ -1,8 +1,8 @@
 export * from './command'
 export * from './config'
 export * from './configWatcher'
+export * from './fetchFromHttpd'
 export * from './fileWatcher'
-export * from './httpd'
 export * from './logExtensionActivated'
 export * from './radCli'
 export * from './views'
