+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './array'
+export * from './httpd'
 export * from './prettify'
